requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
python-dotenv==1.0.0
pytest==8.1.1
requests==2.31.0
pytest-asyncio==0.23.0